requirements.txt
python-dotenv
google-generativeai 
faiss-cpu